refactor: remove namespace, separate internal types
