+++ src/typegenTypeHelpers.ts
@@ -253,34 +253,6 @@ export type NeedsResolver<TypeName extends string, FieldName extends string> = H
     : false
   : false
 
-/**
- * Borrowed from `type-fest`
- * Extract the keys from a type where the value type of the key extends the given `Condition`.
- */
-type ConditionalKeys<Base, Condition> = NonNullable<
-  // Wrap in `NonNullable` to strip away the `undefined` type from the produced union.
-  {
-    // Map through all the keys of the given base type.
-    [Key in keyof Base]: Condition extends Base[Key] // Pick only keys with types extending the given `Condition` type.
-      ? Key // Retain this key since the condition passes.
-      : never // Discard this key since the condition fails.
-
-    // Convert the produced object into a union type of the keys which passed the conditional test.
-  }[keyof Base]
->
-
-/**
- * Taken from `type-fest`
- * Pick keys from the shape that matches the given `Condition`.
- */
-export type ConditionalPick<Base, Condition> = Pick<Base, ConditionalKeys<Base, Condition>>
-
-/**
- * Taken from `type-fest`
- * Get the values of a mapped types
- */
-export type ValueOf<ObjectType, ValueType extends keyof ObjectType = keyof ObjectType> = ObjectType[ValueType]
-
 export type IsFeatureEnabled2<PathPart1 extends string, PathPart2 extends string> = GetGen3<
   'features',
   PathPart1,
@@ -301,121 +273,3 @@ export type Discriminate<
   : Required extends 'required'
   ? Type & { __typename: TypeName }
   : Type & { __typename?: TypeName }
-
-export namespace AbstractTypes {
-  /**
-   * Returns a union of all the type names of the members of an abstract type
-   *
-   * @example
-   *
-   * union D = A | B | C
-   * PossibleTypeNames<'D> // 'A' | 'B' | 'C'
-   */
-  export type PossibleTypeNames<AbstractTypeName extends string> = ValueOf<
-    ConditionalPick<GetGen<'abstractTypeMembers'>, AbstractTypeName>
-  >
-  /**
-   * Returns a union of all the members of an abstract type
-   *
-   * @example
-   * union D = A | B | C
-   * PossibleTypes<'D> // A | B | C
-   */
-  export type PossibleTypes<AbstractTypeName extends string> = RootValue<PossibleTypeNames<AbstractTypeName>>
-
-  /**
-   * Returns a union of all the abstract type names where TypeName is used
-   *
-   * @example
-   * union D = A | B
-   * union E = A
-   * AbstractTypeNames<'A'> // 'D' | 'E'
-   */
-  export type AbstractTypeNames<TypeName extends string> = ConditionalKeys<
-    GetGen<'abstractTypeMembers'>,
-    TypeName
-  >
-
-  /**
-   * Returns whether all the abstract type names where TypeName is used have implemented `resolveType`
-   */
-  export type IsStrategyResolveTypeImplementedInAllAbstractTypes<TypeName extends string> = AbstractTypeNames<
-    TypeName
-  > extends GetGen<'abstractsUsingStrategyResolveType'>
-    ? true
-    : false
-
-  /**
-   * Returns whether all the members of an abstract type have implemented `isTypeOf`
-   */
-  export type IsStrategyIsTypeOfImplementedInAllMembers<AbstractTypeName extends string> = GetGen2<
-    'abstractTypeMembers',
-    AbstractTypeName
-  > extends GetGen<'objectsUsingAbstractStrategyIsTypeOf'>
-    ? true
-    : false
-
-  export type IsTypeOfHandler<TypeName extends string> = (
-    source: PossibleTypes<TypeName>, // typed as never if TypeName is not a member of any abstract type
-    context: GetGen<'context'>,
-    info: GraphQLResolveInfo
-  ) => MaybePromise<boolean>
-
-  /**
-   * Get an object with the `isTypeOf` field if applicable for the given object Type.
-   *
-   * @remarks
-   *
-   * Intersect the result of this with other things to build up the final options for a type def.
-   */
-  // prettier-ignore
-  export type MaybeTypeDefConfigFieldIsTypeOf<TypeName extends string> =
-IsFeatureEnabled2<'abstractTypeStrategies', 'isTypeOf'> extends false // is isTypeOf strategy disabled ?
-? {} // then hide isTypeOf property entirely
-: IsStrategyResolveTypeImplementedInAllAbstractTypes<TypeName> extends true // is resolveType implemented in all abstract types where TypeName is a member?
-  ? { isTypeOf?: IsTypeOfHandler<TypeName> } // then make isTypeOf optional
-  : IsFeatureEnabled2<'abstractTypeStrategies', '__typename'> extends true // is __typename strategy is enabled?
-    ? { isTypeOf?: IsTypeOfHandler<TypeName> } // then make isTypeOf optional
-    : AbstractTypeNames<TypeName> extends never  // is TypeName not part of any abstract type?
-    ? { isTypeOf?: IsTypeOfHandler<TypeName> } // then make isTypeOf optional
-    : { isTypeOf: IsTypeOfHandler<TypeName> } // otherwise, make it required
-
-  /**
-   * Get an object with the `resolveType` field if applicable for the given abstract Type.
-   *
-   * @remarks
-   *
-   * Intersect the result of this with other things to build up the final options for a type def.
-   */
-  export type MaybeTypeDefConfigFieldResolveType<TypeName extends string> = IsFeatureEnabled2<
-    'abstractTypeStrategies',
-    'resolveType'
-  > extends false
-    ? {} // remove field altogether is resolveType strategy is disabled
-    : IsStrategyIsTypeOfImplementedInAllMembers<TypeName> extends true
-    ? {
-        /**
-         * Optionally provide a custom type resolver function. If one is not provided,
-         * the default implementation will call `isTypeOf` on each implementing
-         * Object type.
-         */
-        resolveType?: AbstractTypeResolver<TypeName>
-      } // Make resolveType optional when __typename strategy is enabled
-    : IsFeatureEnabled2<'abstractTypeStrategies', '__typename'> extends true
-    ? {
-        /**
-         * Optionally provide a custom type resolver function. If one is not provided,
-         * the default implementation will call `isTypeOf` on each implementing
-         * Object type.
-         */
-        resolveType?: AbstractTypeResolver<TypeName>
-      }
-    : {
-        /**
-         * Optionally provide a custom type resolver function. If one is not provided,
-         * the default implementation will call `isTypeOf` on each implementing
-         * Object type.
-         */
-        resolveType: AbstractTypeResolver<TypeName>
-      }
-}
